% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_curve.R
\name{decision_label_table}
\alias{decision_label_table}
\title{Decision Label Table}
\usage{
decision_label_table(
  dat,
  decision_cols,
  decision_labels,
  filename = "decision_label_table.csv"
)
}
\arguments{
\item{dat}{Properly formatted data set, limited to a single outcome.}

\item{decision_cols}{String vector containing the column names with the analytical decisions.}

\item{decision_labels}{String vector with the names of the analytical decisions, if not the column names themselves.}

\item{filename}{File name to save the table to. Must end in '.csv'. Set to \code{NULL} to not export to file.}
}
\description{
This produces a table showing the correspondence between the numeric values of the analytical decisions and their labels.
}
\details{
Output is a \code{ggplot2} graph, so graphical customization should be possible after the fact if desired.
}
\examples{

# Example data
data(cholera_dat)
spec_curve(cholera_dat[cholera_dat$outcome == 3,],
  decision_cols = c('analytical_decision_1','analytical_decision_2','analytical_decision_3'),
  decision_labels = c('Specification','Bandwidth','Controls'),
  orig_spec_values = c(1,1,1))

}
