% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_curve.R
\name{spec_curve}
\alias{spec_curve}
\title{Specification Curve for R2E-formatted Replication Mixes}
\usage{
spec_curve(
  spec_data,
  decision_cols,
  beta = "beta",
  se = "se",
  decision_labels = NULL,
  orig_beta_se = NULL,
  orig_spec_values = NULL,
  left_box = 50,
  add_hline_at = 0,
  skip_theme = FALSE,
  font_size = 8,
  color_text_labels = NULL,
  height_ratio = c(4, 1),
  return_separate = FALSE
)
}
\arguments{
\item{decision_cols}{String vector containing the column names with the analytical decisions.}

\item{beta}{String variable indicating the column with the effect of interest.}

\item{se}{String variable indicating the column with the standard error.}

\item{decision_labels}{String vector with the names of the analytical decisions, if not the column names themselves.}

\item{orig_beta_se}{A two-element vector containing the original specification's beta and SE values. To be used if the original specification is not in the data. Will assign analytical decision values of 0 unless \code{orig_spec_values} is also specified.}

\item{orig_spec_values}{A vector with the same length as \code{decision_cols} with the values that correspond to the original analysis.}

\item{left_box}{Width of y-axis title area}

\item{add_hline_at}{Add a horizontal line at this value. Set to \code{NULL} to omit.}

\item{skip_theme}{Don't apply the package theme, instead use \code{ggplot2} defaults.}

\item{font_size}{Set font size in points for the variant box at bottom.}

\item{color_text_labels}{Use a gradient color scale for the text labels (requires numeric analytic decision columns). Set to \code{TRUE} to include, or to a \code{ggplot2::scale_color} function to specify your own color scale.}

\item{height_ratio}{The ratio of the top plot to the bottom plot.}

\item{return_separate}{Return two ggplots separately in a list so you can combine them yourself.}

\item{dat}{Properly formatted data set, limited to a single outcome.}
}
\description{
This produces a specification curve if given data matching the R2E format.
}
\details{
Output is a \code{ggplot2} graph, so graphical customization should be possible after the fact if desired.
}
\examples{

# Example data
data(cholera_dat)
spec_curve(cholera_dat[cholera_dat$outcome == 3,],
  decision_cols = c('analytical_decision_1','analytical_decision_2','analytical_decision_3'),
  decision_labels = c('Specification','Bandwidth','Controls'),
  orig_spec_values = c(1,1,1))

}
